refactor: move AuthMiddleware dependencies to a protocols file
